refactor: price api
