+++ packages/web3-providers/src/coingecko/helper.ts
@@ -1,11 +1,11 @@
+import { createLookupTableResolver } from '@masknet/shared-base'
 import { ChainId } from '@masknet/web3-shared-evm'
 
-const NETWORK_NAME_MAP: { [key in string]: ChainId } = {
-    ethereum: ChainId.Mainnet,
-    'binance-smart-chain': ChainId.BSCT,
-    'polygon-pos': ChainId.Matic,
-}
-
-export function resolveChain(key: string) {
-    return NETWORK_NAME_MAP[key]
-}
+export const resolveChain = createLookupTableResolver<string, ChainId | undefined>(
+    {
+        ethereum: ChainId.Mainnet,
+        'binance-smart-chain': ChainId.BSCT,
+        'polygon-pos': ChainId.Matic,
+    },
+    undefined,
+)
